Add more info to the 'Conflicts' section
Introduce 'git rebase --show-current-patch' and REBASE_HEAD, and
mention how 'git checkout [--ours | --theirs]' is interpreted in the
context of a rebase merge conflict.

For bonus points, mention that you can recreate the merge conflict with
'git checkout --merge .' so that the conflict markers are clearly labeled
as 'ours' and 'theirs'.